requirements.txt
python-telegram-bot
python-dotenv
openai
langchain
langchain-openai
langchain-text-splitters
langchain-community
faiss-cpu
requests
beautifulsoup4
langdetect